# api.yaml
openapi: 3.0.3
info:
  title: AI Assistants Catalog
  description: Каталог AI-ассистентов с React-клиентом и мокируемым LLM-провайдером
  version: "1.0.0"

servers:
  - url: http://localhost:8080

tags:
  - name: System
  - name: Auth
  - name: Categories
  - name: Assistants
  - name: Runs

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    CategoryIdPath:
      name: categoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Идентификатор категории
    AssistantIdPath:
      name: assistantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Идентификатор ассистента
    RunIdPath:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Идентификатор запуска

  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INVALID_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - EMAIL_ALREADY_EXISTS
                - CATEGORY_NOT_FOUND
                - ASSISTANT_NOT_FOUND
                - ASSISTANT_INACTIVE
                - LLM_PROVIDER_ERROR
                - INTERNAL_ERROR
            message:
              type: string
      example:
        error:
          code: INVALID_REQUEST
          message: invalid request

    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0

    Role:
      type: string
      enum: [admin, user]

    Token:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT для заголовка Authorization Bearer
        user:
          $ref: "#/components/schemas/User"

    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
          nullable: true

    AuthCredentialsIn:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 72
          description: Минимум 8 символов, хотя бы одна буква и хотя бы одна цифра или спецсимвол

    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Еда
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true

    CategoryCreateIn:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true

    Assistant:
      type: object
      required: [id, categoryId, name, description, model, isActive, isFavorite]
      properties:
        id:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        categoryName:
          type: string
          nullable: true
        name:
          type: string
          example: Повар
        description:
          type: string
          example: Составляет подробный рецепт по списку ингредиентов
        model:
          type: string
          example: mock-smart
        systemPrompt:
          type: string
          nullable: true
          description: Может быть скрыт от обычного пользователя по решению кандидата
        exampleUserPrompt:
          type: string
          nullable: true
          example: курица, рис, томаты, сыр
        isActive:
          type: boolean
        isFavorite:
          type: boolean
          description: Ассистент добавлен в избранное текущим пользователем
        createdAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true

    AssistantCreateIn:
      type: object
      required: [categoryId, name, description, model, systemPrompt]
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        systemPrompt:
          type: string
          minLength: 1
        exampleUserPrompt:
          type: string
          nullable: true
        isActive:
          type: boolean
          default: true

    AssistantUpdateIn:
      type: object
      required: [categoryId, name, description, model, systemPrompt, isActive]
      properties:
        categoryId:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        model:
          type: string
          minLength: 1
        systemPrompt:
          type: string
          minLength: 1
        exampleUserPrompt:
          type: string
          nullable: true
        isActive:
          type: boolean

    RunStatus:
      type: string
      enum: [pending, success, failed]
      description: >
        pending — запуск создан, но ответ ещё не получен;
        success — запуск завершился успешно;
        failed — запуск завершился ошибкой.

    RunFeedbackRating:
      type: integer
      format: int32
      enum: [-1, 1]
      description: -1 — дизлайк, 1 — лайк.

    RunFeedbackUpsertIn:
      type: object
      required: [rating]
      properties:
        rating:
          $ref: "#/components/schemas/RunFeedbackRating"

    AssistantRun:
      type: object
      required: [id, assistantId, userId, model, userPrompt, status]
      properties:
        id:
          type: string
          format: uuid
        assistantId:
          type: string
          format: uuid
        assistantName:
          type: string
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        categoryName:
          type: string
          nullable: true
        userId:
          type: string
          format: uuid
        model:
          type: string
        userPrompt:
          type: string
        output:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/RunStatus"
        error:
          type: string
          nullable: true
        feedbackRating:
          allOf:
            - $ref: "#/components/schemas/RunFeedbackRating"
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true

    AssistantRunCreateIn:
      type: object
      required: [userPrompt]
      properties:
        userPrompt:
          type: string
          minLength: 1

    AssistantRunStreamDelta:
      type: object
      required: [delta]
      properties:
        delta:
          type: string
          description: Очередной фрагмент ответа LLM.

    AssistantRunStreamFailure:
      type: object
      required: [run, error]
      properties:
        run:
          $ref: "#/components/schemas/AssistantRun"
        error:
          $ref: "#/components/schemas/ErrorResponse/properties/error"

paths:
  /_info:
    get:
      tags: [System]
      summary: Healthcheck
      responses:
        "200":
          description: OK

  /auth/register:
    post:
      tags: [Auth]
      summary: Зарегистрировать пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentialsIn"
      responses:
        "201":
          description: JWT и новый пользователь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Пользователь с такой почтой уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      tags: [Auth]
      summary: Войти по email и паролю
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthCredentialsIn"
      responses:
        "200":
          description: JWT и пользователь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dummyLogin:
    post:
      tags: [Auth]
      summary: Получить тестовый JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  $ref: "#/components/schemas/Role"
      responses:
        "200":
          description: JWT и пользователь
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Token"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories:
    get:
      tags: [Categories]
      summary: Список категорий
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Список категорий
          content:
            application/json:
              schema:
                type: object
                required: [categories]
                properties:
                  categories:
                    type: array
                    items:
                      $ref: "#/components/schemas/Category"
        "401":
          description: Нет авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Categories]
      summary: Создать категорию
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryCreateIn"
      responses:
        "201":
          description: Категория создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assistants:
    get:
      tags: [Assistants]
      summary: Список ассистентов
      security:
        - BearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
          description: Фильтр по категории
        - name: q
          in: query
          schema:
            type: string
          description: Поиск по названию и описанию
        - name: includeInactive
          in: query
          schema:
            type: boolean
            default: false
          description: Фильтр "Показать неактивных ассистентов"; доступен администратору
        - name: favoriteOnly
          in: query
          schema:
            type: boolean
            default: false
          description: Показывать только избранных ассистентов текущего пользователя
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Список ассистентов
          content:
            application/json:
              schema:
                type: object
                required: [assistants, pagination]
                properties:
                  assistants:
                    type: array
                    items:
                      $ref: "#/components/schemas/Assistant"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          description: Нет авторизации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Assistants]
      summary: Создать ассистента
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantCreateIn"
      responses:
        "201":
          description: Ассистент создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assistants/{assistantId}:
    get:
      tags: [Assistants]
      summary: Получить ассистента
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssistantIdPath"
      responses:
        "200":
          description: Ассистент
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "404":
          description: Ассистент не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags: [Assistants]
      summary: Обновить ассистента
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssistantIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantUpdateIn"
      responses:
        "200":
          description: Ассистент обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ассистент не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assistants/{assistantId}/favorite:
    put:
      tags: [Assistants]
      summary: Добавить ассистента в избранное
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssistantIdPath"
      responses:
        "204":
          description: Ассистент добавлен в избранное
        "400":
          description: Некорректный идентификатор
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Требуется аутентификация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ассистент не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Assistants]
      summary: Убрать ассистента из избранного
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssistantIdPath"
      responses:
        "204":
          description: Ассистент убран из избранного
        "400":
          description: Некорректный идентификатор
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Требуется аутентификация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ассистент не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assistants/{assistantId}/run:
    post:
      tags: [Runs]
      summary: Запустить ассистента
      description: Backend использует системный промпт ассистента и пользовательский промпт из тела запроса.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssistantIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantRunCreateIn"
      responses:
        "201":
          description: Запуск создан или завершён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantRun"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Ассистент выключен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Ошибка LLM-провайдера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /assistants/{assistantId}/run/stream:
    post:
      tags: [Runs]
      summary: Запустить ассистента со streaming-ответом
      description: >
        Backend создаёт pending-запуск, отдаёт прогресс генерации через Server-Sent Events
        и после завершения сохраняет финальный output в историю. События:
        `run` — созданный pending-запуск (`AssistantRun`);
        `delta` — фрагмент ответа (`AssistantRunStreamDelta`);
        `done` — финальный успешный запуск (`AssistantRun`);
        `failed` — финальный failed-запуск и ошибка (`AssistantRunStreamFailure`).
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AssistantIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantRunCreateIn"
      responses:
        "200":
          description: SSE-поток прогресса генерации
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                stream:
                  value: |
                    event: run
                    data: {"id":"...","assistantId":"...","userId":"...","model":"openai/gpt-4o-mini","userPrompt":"hi","status":"pending"}

                    event: delta
                    data: {"delta":"Hello"}

                    event: done
                    data: {"id":"...","assistantId":"...","userId":"...","model":"openai/gpt-4o-mini","userPrompt":"hi","output":"Hello","status":"success"}
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Требуется аутентификация
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Ассистент не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Ассистент выключен
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Внутренняя ошибка до старта SSE-потока
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Ошибка LLM-провайдера до старта SSE-потока
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /runs/{runId}/feedback:
    put:
      tags: [Runs]
      summary: Оценить или сменить оценку ответа ассистента
      description: Пользователь может оценивать только свои запуски.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RunIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunFeedbackUpsertIn"
      responses:
        "200":
          description: Запуск с обновлённой оценкой
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantRun"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Запуск принадлежит другому пользователю
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Запуск не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /runs/my:
    get:
      tags: [Runs]
      summary: История запусков текущего пользователя
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: История запусков
          content:
            application/json:
              schema:
                type: object
                required: [runs, pagination]
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssistantRun"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

  /admin/runs:
    get:
      tags: [Runs]
      summary: Все запуски ассистентов для администратора
      security:
        - BearerAuth: []
      parameters:
        - name: assistantId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Список запусков
          content:
            application/json:
              schema:
                type: object
                required: [runs, pagination]
                properties:
                  runs:
                    type: array
                    items:
                      $ref: "#/components/schemas/AssistantRun"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"